Move std_links to its own module.
This code is getting long enough that it will help to organize it separately.
