+++ mdbook-spec/src/main.rs
@@ -6,11 +6,11 @@ use once_cell::sync::Lazy;
 use regex::{Captures, Regex};
 use semver::{Version, VersionReq};
 use std::collections::BTreeMap;
-use std::fmt::Write as _;
-use std::fs;
-use std::io::{self, Write as _};
+use std::io;
 use std::path::PathBuf;
-use std::process::{self, Command};
+use std::process;
+
+mod std_links;
 
 /// The Regex for rules like `r[foo]`.
 static RULE_RE: Lazy<Regex> = Lazy::new(|| Regex::new(r"(?m)^r\[([^]]+)]$").unwrap());
@@ -20,31 +20,6 @@ static RULE_RE: Lazy<Regex> = Lazy::new(|| Regex::new(r"(?m)^r\[([^]]+)]$").unwr
 static ADMONITION_RE: Lazy<Regex> = Lazy::new(|| {
     Regex::new(r"(?m)^ *> \[!(?<admon>[^]]+)\]\n(?<blockquote>(?: *> .*\n)+)").unwrap()
 });
-
-/// A markdown link (without the brackets) that might possibly be a link to
-/// the standard library using rustdoc's intra-doc notation.
-const STD_LINK: &str = r"(?: [a-z]+@ )?
-                         (?: std|core|alloc|proc_macro|test )
-                         (?: ::[A-Za-z0-9_!:<>{}()\[\]]+ )?";
-
-/// The Regex for a markdown link that might be a link to the standard library.
-static STD_LINK_RE: Lazy<Regex> = Lazy::new(|| {
-    Regex::new(&format!(
-        r"(?x)
-            (?:
-                ( \[`[^`]+`\] ) \( ({STD_LINK}) \)
-            )
-            | (?:
-                ( \[`{STD_LINK}`\] )
-            )
-         "
-    ))
-    .unwrap()
-});
-
-/// The Regex used to extract the std links from the HTML generated by rustdoc.
-static STD_LINK_EXTRACT_RE: Lazy<Regex> =
-    Lazy::new(|| Regex::new(r#"<li><a [^>]*href="(https://doc.rust-lang.org/[^"]+)""#).unwrap());
 
 fn main() {
     let mut args = std::env::args().skip(1);
@@ -184,119 +159,6 @@ impl Spec {
             })
             .to_string()
     }
-
-    /// Converts links to the standard library to the online documentation in
-    /// a fashion similar to rustdoc intra-doc links.
-    fn std_links(&self, chapter: &Chapter) -> String {
-        // This is very hacky, but should work well enough.
-        //
-        // Collect all standard library links.
-        //
-        // links are tuples of ("[`std::foo`]", None) for links without dest,
-        // or ("[`foo`]", "std::foo") with a dest.
-        let mut links: Vec<_> = STD_LINK_RE
-            .captures_iter(&chapter.content)
-            .map(|cap| {
-                if let Some(no_dest) = cap.get(3) {
-                    (no_dest.as_str(), None)
-                } else {
-                    (
-                        cap.get(1).unwrap().as_str(),
-                        Some(cap.get(2).unwrap().as_str()),
-                    )
-                }
-            })
-            .collect();
-        if links.is_empty() {
-            return chapter.content.clone();
-        }
-        links.sort();
-        links.dedup();
-
-        // Write a Rust source file to use with rustdoc to generate intra-doc links.
-        let tmp = tempfile::TempDir::with_prefix("mdbook-spec-").unwrap();
-        let src_path = tmp.path().join("a.rs");
-        // Allow redundant since there could some in-scope things that are
-        // technically not necessary, but we don't care about (like
-        // [`Option`](std::option::Option)).
-        let mut src = format!(
-            "#![deny(rustdoc::broken_intra_doc_links)]\n\
-             #![allow(rustdoc::redundant_explicit_links)]\n"
-        );
-        for (link, dest) in &links {
-            write!(src, "//! - {link}").unwrap();
-            if let Some(dest) = dest {
-                write!(src, "({})", dest).unwrap();
-            }
-            src.push('\n');
-        }
-        writeln!(
-            src,
-            "extern crate alloc;\n\
-             extern crate proc_macro;\n\
-             extern crate test;\n"
-        )
-        .unwrap();
-        fs::write(&src_path, &src).unwrap();
-        let output = Command::new("rustdoc")
-            .arg("--edition=2021")
-            .arg(&src_path)
-            .current_dir(tmp.path())
-            .output()
-            .expect("rustdoc installed");
-        if !output.status.success() {
-            eprintln!(
-                "error: failed to extract std links ({:?}) in chapter {} ({:?})\n",
-                output.status,
-                chapter.name,
-                chapter.source_path.as_ref().unwrap()
-            );
-            io::stderr().write_all(&output.stderr).unwrap();
-            process::exit(1);
-        }
-
-        // Extract the links from the generated html.
-        let generated =
-            fs::read_to_string(tmp.path().join("doc/a/index.html")).expect("index.html generated");
-        let urls: Vec<_> = STD_LINK_EXTRACT_RE
-            .captures_iter(&generated)
-            .map(|cap| cap.get(1).unwrap().as_str())
-            .collect();
-        if urls.len() != links.len() {
-            eprintln!(
-                "error: expected rustdoc to generate {} links, but found {} in chapter {} ({:?})",
-                links.len(),
-                urls.len(),
-                chapter.name,
-                chapter.source_path.as_ref().unwrap()
-            );
-            process::exit(1);
-        }
-
-        // Replace any disambiguated links with just the disambiguation.
-        let mut output = STD_LINK_RE
-            .replace_all(&chapter.content, |caps: &Captures| {
-                if let Some(dest) = caps.get(2) {
-                    // Replace destination parenthesis with a link definition (square brackets).
-                    format!("{}[{}]", &caps[1], dest.as_str())
-                } else {
-                    caps[0].to_string()
-                }
-            })
-            .to_string();
-
-        // Append the link definitions to the bottom of the chapter.
-        write!(output, "\n").unwrap();
-        for ((link, dest), url) in links.iter().zip(urls) {
-            if let Some(dest) = dest {
-                write!(output, "[{dest}]: {url}\n").unwrap();
-            } else {
-                write!(output, "{link}: {url}\n").unwrap();
-            }
-        }
-
-        output
-    }
 }
 
 impl Preprocessor for Spec {
@@ -315,7 +177,7 @@ impl Preprocessor for Spec {
             }
             ch.content = self.rule_definitions(&ch, &mut found_rules);
             ch.content = self.admonitions(&ch);
-            ch.content = self.std_links(&ch);
+            ch.content = std_links::std_links(&ch);
         }
         for section in &mut book.sections {
             let BookItem::Chapter(ch) = section else {
